Relocated LatLonOffset and related utility to dataClasses
This commit moves `LatLonOffset` and the `convertToLatLonOffsetList` function from the `extrasForUI` package to the `dataClasses` package to better reflect their role as data structures.

**Key Changes:**

*   **`LatLonOffSet.kt`**:
    *   Renamed and moved from `app/src/main/java/com/example/trackpro/extrasForUI/CanvasMap.kt` to `app/src/main/java/com/example/trackpro/dataClasses/LatLonOffSet.kt`.
    *   Updated the package declaration to `com.example.trackpro.dataClasses`.
    *   Simplified the `convertToLatLonOffsetList` function signature by removing the redundant full classpath for `RawGPSData`.
*   **Package Imports**:
    *   Updated imports from `com.example.trackpro.extrasForUI.LatLonOffset` to `com.example.trackpro.dataClasses.LatLonOffset` across the following files:
        *   `TimeAttackViewModel.kt`
        *   `DragTimeCalculation.kt`
        *   `DragScreenListItem.kt`
        *   `TrackBuilderScreen.kt`
        *   `TimeAttackScreen.kt`
*   **`TrackBuilderScreen.kt`**: Removed unused `androidx.compose.foundation.Canvas` import.
