--- app/src/main/java/com/example/trackpro/extrasForUI/CanvasMap.kt
@@ -1,9 +1,0 @@
-package com.example.trackpro.extrasForUI
-
-data class LatLonOffset(val lat: Double, val lon: Double)
-
-fun convertToLatLonOffsetList(data: List<com.example.trackpro.dataClasses.RawGPSData>): List<LatLonOffset> {
-    return data.map { LatLonOffset(lat = it.latitude, lon = it.longitude) }
-}
-
-
